requirements.txt
openai
pyheif
selenium
webdriver-manager
pillow
psycopg2
requests
python-dotenv
tabulate